before precaching data

+++ core/region.go
@@ -2,6 +2,6 @@ package core
 
 // Region representing regions
 type Region struct {
-	ID   int    `json:"id"`
+	ID   int64  `json:"id"`
 	Name string `json:"name"`
 }
